Add language file validation test
